fix: add src dir
Apparently jdtls wants code to be in a directory called 'src,'
otherwise it won't be able to resolve packages.
