+++ src/tsgui/InstrReg.java
